+++ modulo3/projeto-labEddit/labeddit/src/Services/posts.js
@@ -1,0 +1,6 @@
+import axios from 'axios'
+import { BASE_URL } from '../constants/urls'
+
+export const createPost = (body, ) => {
+
+}
