+++ modulo3/projeto-labEddit/labeddit/src/components/Loading/Loading.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import LoadinGifPage from '../../assets/loadingPage.gif'
+
+const Loading = () => {
+  return (
+    <div>
+        <img src={LoadinGifPage} alt='Gif de carregando' />
+    </div>
+  )
+}
+
+export default Loading
