+++ modulo3/projeto-labEddit/labeddit/src/components/Loading/LoadingStyled.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const LoadingContainer = styled.section`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`
+
+
+
